third_party/blink/web_tests/transforms/2d/computed-style-origin-expected.txt

This tests computed style values for transform-origin and perspective-origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test1Style.webkitTransformOrigin is '10px 10px'
PASS test2Style.webkitTransformOrigin is '10px 10px'
PASS test3Style.webkitTransformOrigin is '0px 20px'
PASS test4Style.webkitTransformOrigin is '5px 7px'
PASS test5Style.webkitTransformOrigin is '15px 15px'
PASS test1Style.webkitPerspectiveOrigin is '10px 10px'
PASS test2Style.webkitPerspectiveOrigin is '10px 10px'
PASS test3Style.webkitPerspectiveOrigin is '0px 20px'
PASS test4Style.webkitPerspectiveOrigin is '5px 7px'
PASS test5Style.webkitPerspectiveOrigin is '15px 15px'
PASS test6Style.webkitPerspectiveOrigin is '50px 50px'


PASS successfullyParsed is true

TEST COMPLETE
